+++ RedshiftLibrary/src/lib/method/linemodelsolve.cpp
@@ -203,19 +203,6 @@ std::shared_ptr<CSolveResult> CLineModelSolve::compute() {
   return lmSolveResult;
 }
 
-COperatorPdfz CLineModelSolve::initializePdfz(Int32 maxPeakPerWindow,
-                                              Int32 peakSeparation,
-                                              Int32 cutThreshold,
-                                              Int32 extremaCount) const {
-  COperatorPdfz pdfz(m_opt_pdfcombination, peakSeparation, cutThreshold,
-                     extremaCount, m_zLogSampling,
-                     "SPE",           // Id_prefix
-                     false,           // do not allow extrema at border
-                     maxPeakPerWindow // one peak/window only
-  );
-  return pdfz;
-}
-
 void CLineModelSolve::GetZpriorsOptions(
     bool &zPriorStrongLinePresence, bool &zPriorHaStrongestLine,
     bool &zPriorNLineSNR, Float64 &opt_nlines_snr_penalization_factor,