+++ CHANGES.md
@@ -9,7 +9,8 @@
 ### Removed
 * [9830] : Remove CFITSIO dependency
 ### Fixed
-* [9615] fix unintentional numerous warnings when continuum is null and fitting absorption lines
+* [9615] : Fix unintentional numerous warnings when continuum is null and fitting absorption lines
+* [9616] : Fix linemeas "empty line element list to fit"
 
 ## 1.8.0 (2025-06-06)
 ### Added