+++ CHANGES.md
@@ -5,6 +5,7 @@
 * [9715] : Add goodness of fit for template fitting with FFT processing algorithm
 ### Changed
 * [9830] : Switch to lBFGSb v0.4.0
+* [9355] : Use lBFGSb inverse Hessian approximation to compute linemeas uncertainties
 ### Deprecated
 ### Removed
 * [9830] : Remove CFITSIO dependency