+++ CHANGES.md
@@ -15,6 +15,7 @@
 * [9615] : Fix unintentional numerous warnings when continuum is null and fitting absorption lines
 * [9616] : Fix linemeas "empty line element list to fit"
 * [9836] : Fix template fitting 2nd pass evidence saved in ResultStore
+* [9885] : Fix template fitting 2nd pass with candidate at igm border
 
 ## 1.8.0 (2025-06-06)
 ### Added